Add concept of favorite trip
Adds the concept of favorite trip to the app. The trip that will be displayed at the top of the trips page will be the persons favorite trip. That can either be the trip the person is currently on that they are adding stuff too or it can be the trip they enjoy the most that they want easy access too.