Move Navigation View to Out From List View
Move the Navigation view for trips out from the trips list view to the trips page as a whole. Causes whole page to reset to trip detail page instead of only half of the page. Also allows the navigation for favorite trip section and for the add new trip section"