Add skeleton to trip detail page
Adds the three main sections. Trip information will contain the basic info including start and end date, people included in the trip, and location. The Activieies section will contain a list of activities included on the trip. The Photos section will have the photos of the trip laid out in a grid section. This also places the edit button and name of the trip in the navigation bar of the view.