Add view for activities list
List for activities appears on the trip detail page. It allows users to quickly see all of the activities they have linked to a given trip. In the future I will add a Plus icon next to the activities section title so users can easily attatch new activities to a trip