Fix display of activities list
Fixes core styling issues on the trip detail page. Also implements an activity detail page and routing to it from the activities list view.

As a side note fixed a quirky bug on the my trips page where the trip toggle button was initially loading as future trips even though the future trips list was the one loading onto the screen