Add navigation logic to explore page
Add navigation to the trip list and users list on the explore page. The trip tiles link to a trip detail page for that trip and the user tiles link to a profile view for that user. All of the navigation stays within the navigation view of the explore page so the user can follow the chain as long as they want and then go back and keep exploring