Prevent annotations from clustering
Prevent annotations on the map from clustering. I think initially I don't want annotations to cluster because I am not worried about crowding on the map necissarily. Right now the map is mainly for visual purposes anyways.

The only issue currently is that there is no way to tell if there are two markers in the exact same place.